requirements.txt
cloudpickle==2.2.0
dateparser==1.2.0
dgl==1.1.2+cu118
elasticsearch==8.5.0
en_core_web_trf==3.7.3
gensim==3.8.1
joblib==1.2.0
langdetect==1.0.7
matplotlib==3.7.1
memory_profiler==0.61.0
nltk==3.7
nose==1.3.7
num2words==0.5.12
numpy==1.23.4
openai==0.27.2
pandas==1.5.1
psutil==5.9.4
pycountry==22.3.5
pytest==7.2.2
python_dateutil==2.8.2
regex==2022.3.2
reporters_db==3.2.32
Requests==2.31.0
scikit_learn==0.24.0
scipy==1.12.0
sentence_transformers.egg==info
spacy==3.7.2
tiktoken==0.3.2
torch==2.0.0+cu118
torchmetrics==0.11.4
tqdm==4.64.1
transformers==4.27.3
Unidecode==1.3.6
zahlwort2num==0.4.2